TJNP: - Moved TJShow's FileServer and Authorizer classes over to TJNP (class WebServer/Authorizer) - Removed the general tjnp.h include file; users of TJNP should now only include the functionality they need. - Pattern is not a template class anymore, but instead only handles wchar_t strings.
TJShared:
- File::PathSeparator => File::GetPathSeparator


git-svn-id: https://svn.pixelspark.nl/svn/corespark/trunk@550 c9a305e2-9f8a-4be3-8c95-b0e5bab41dc9

+++ TJShared/include/tjutil.h
@@ -178,7 +178,7 @@ namespace tj {
 		}
 
 		inline std::string Mbs(const String& ws) {
-			unsigned int n = ws.length();
+			unsigned int n = (unsigned int)ws.length();
 			char* buf  = new char[n+1];
 			
 			#ifdef TJ_OS_MAC
@@ -196,7 +196,7 @@ namespace tj {
 		}
 
 		inline String Wcs(const std::string& ws) {
-			unsigned int n = ws.length();
+			unsigned int n = (unsigned int)ws.length();
 			
 			#ifdef TJ_OS_WIN
 				wchar_t* buf  = new wchar_t[n+1];
